fix(firebase): esbuild config
